+++ src/game.ts
@@ -1,6 +1,9 @@
 import * as Phaser from 'phaser';
 
 import { GAME_SZ } from './constants';
+import { HomeScreen } from './scenes/HomeScreen';
+import { Level } from './scenes/Level';
+import { LoadAssets } from './scenes/LoadAssets';
 
 const config: Phaser.Types.Core.GameConfig = {
     parent: 'game',
@@ -19,6 +22,14 @@ const config: Phaser.Types.Core.GameConfig = {
     },
     scale: {
         mode: Phaser.Scale.FIT
+    },
+    physics:{
+        default: "arcade",
+        arcade: { 
+            debug: true
+            //gravity: {y: 5000},
+            //tileBias: 96
+        }
     }
 };
 
@@ -26,6 +37,10 @@ export class KTGame extends Phaser.Game {
 
     constructor(config: Phaser.Types.Core.GameConfig) {
         super(config);
+        //scenes
+        this.scene.add("LoadAssets", LoadAssets, true)
+        this.scene.add("HomeScreen", HomeScreen, false);
+        this.scene.add("Level", Level, false);
     }
 
 }