+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Version 1
+## Version 1 Inital Setup
 * Setup inital project structure
 * Pulled down KT Starter
 * Created Source Manager
@@ -6,3 +6,9 @@
 * Added npm commands for updating major, minor, and patch updates
 * Pulled inital assets
 
+## Version 2 Scene Essentials
+* 2.0.0: Pulled utilities
+* 2.1.0: Setup Home Screen and Level scenes and made test scene in tiled
+* 2.2.0: Got level properly loaded in and created and added Load assets scene.
+
+## Version 3